chore: refactored the mapping logic
